hexagon: cleanup common header used in host/npu
